src/tensorrt/tensorrt-6.0.1.5/samples/python/fc_plugin_caffe_mnist/requirements.txt
Pillow
pycuda
numpy
argparse
